Fix minor cast warning.

+++ src/GUI/selformats.h
@@ -16,7 +16,7 @@ public:
 	{
 		//Cannot get SetSize() to work here so use fixed item height
 		size_t height = 5 + 15 * AllFormatsArray.GetCount();
-		size_t maxheight = 0.8*wxSystemSettings::GetMetric(wxSYS_SCREEN_Y);
+		size_t maxheight = static_cast<size_t> (0.8*wxSystemSettings::GetMetric(wxSYS_SCREEN_Y));
 		if(height > maxheight)
 			height = maxheight;
 		pList = new wxCheckListBox(this, wxID_ANY, wxDefaultPosition,
